Moved multiverse perms to InstancePlayer to affect editing worlds too.
